CRUD Commands implementation and GlobalExceptionHandler implementation

+++ src/User.Example.Application/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -2,8 +2,9 @@
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Collections.Generic;
 
-namespace User.Example.Application.Commands.UserCmd
+namespace User.Example.Application.Commands.DeleteUser
 {
     public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, bool>
     {
@@ -15,7 +16,15 @@ namespace User.Example.Application.Commands.UserCmd
 
         public async Task<bool> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
+            if (!UserExist(request.Identification))
+                throw new KeyNotFoundException($"User with id {request.Identification} does not exist.");
+
             return _userRepository.Delete(request.Identification);
+        }
+
+        private bool UserExist(string identification)
+        {
+            return _userRepository.GetById(identification) != null;
         }
     }
 }
